% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision-tree-example.R
\docType{data}
\name{iris}
\alias{iris}
\title{Decision tree example
Trying out rpart}
\format{An object of class \code{data.frame} with 150 rows and 5 columns.}
\usage{
iris
}
\description{
Decision tree example
Trying out rpart
}
\keyword{datasets}

